foglalasEStorles

+++ server/public/js/orvos/torol.js
@@ -1,19 +1,19 @@
 function torol(id) {
-    fetch(`/torold/${id}`, {
-        method: 'DELETE',
-        headers: {
-                    'Content-Type': 'application/json',
-                },
-                
-    })
-    .then(response => response.json())
-    .then(data => {
-        alert(data.msg);
-        
-        location.reload();
-    })
-    .catch(error => {
-        console.error('Hiba történt:', error);
-        alert('Hiba történt a törlés során.');
-    });
+    if (confirm("Biztosan törölni szeretnéd ezt az orvost?")) {
+        fetch(`/torold/${id}`, {
+            method: 'DELETE',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+        })
+        .then(response => response.json())
+        .then(data => {
+            alert(data.msg);
+            location.reload();
+        })
+        .catch(error => {
+            console.error('Hiba történt:', error);
+            alert('Hiba történt a törlés során.');
+        });
+    }
 }
